+++ package.json
@@ -13,12 +13,14 @@
     "@react-navigation/bottom-tabs": "^6.3.1",
     "@react-navigation/native": "^6.0.8",
     "@react-navigation/native-stack": "^6.5.0",
+    "axios": "^0.26.1",
     "metro-transform-worker": "^0.69.0",
     "react": "17.0.2",
     "react-native": "0.67.2",
     "react-native-safe-area-context": "^3.4.0",
     "react-native-screens": "^3.11.1",
-    "react-native-vector-icons": "^9.1.0"
+    "react-native-vector-icons": "^9.1.0",
+    "rxjs": "^7.5.5"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
