+++ src/pages/Settings.tsx
@@ -1,10 +1,6 @@
-import React from "react";
-import {View} from "react-native";
+import React from 'react';
+import { View } from 'react-native';
 
-export default function Settings({navigation}:any) {
-    return(
-        <View>
-
-        </View>
-    )
+export default function Settings({ navigation }: any) {
+  return <View />;
 }
